Add checks on Openstack side also
Now images are checked on the openstack side also before uploading.
If the currently downloaded image is already on the server then it is
not uploaded otherwise it will be. Local storage of files is updated as
before.

Also included with this commit are small improvements to the code such
as more printing outputs, cleaner checksum verification, qemu-img no
longer assuming everything is a qcow2, ability to enable verbose
openstack debugging and hashing a file is now a function which can be
called from other places if needed. README.md has been updated to
reflect these changes and it also now features instructions how to set
up a virtual environment and install the required packages for it.