Add properties to JSON and improve checksums
Now any properties can be defined in the JSON.
Checksum validation now takes into account the checksum on the
openstack side. Checksum warnings are better also.
Security groups are now assumed to exist but created if not.
Progress bars are hidden if there's no tty present in order to keep log
files sensible.
Deleting old images is moved to its own function to clean up the code.
Now an error is thrown if cloud and/or network is not defined in the
environment.